agent/requirements.txt
vision-agents[getstream,gemini]
python-dotenv
fastapi
uvicorn
